added tests for upcomingScreenings and fixed upcomingScreenings function.
